Squashed 'seims/src/ccgl/' changes from ff5bf61cb..926c2f39a
926c2f39a Don't test mongodb if connect failed
376b49b0a handle the difference bewteen localtime and utctime
62edf9c02 test on msvc+intel c++ compiler
7c095aab6 Use emplace and emplace_back rather than insert and push_back whenever possible
ccd417a13 Merge branch 'master' of github.com:crazyzlj/CCGL
c3c8d4960 add include for macos
92d099c85 little bug fixed when figure out whether rebuild data is needed
a35dd2756 delete tm pointer

git-subtree-dir: seims/src/ccgl
git-subtree-split: 926c2f39a8c54bb4cca078237c709bc1c75c5043

+++ test/raster/test_2d-pos_incst-mask-nopos-ext.cpp
@@ -217,20 +217,22 @@ TEST_P(clsRasterDataTestMultiPosIncstMaskNoPosExt, RasterIO) {
 #ifdef USE_MONGODB
     /** MongoDB I/O test **/
     MongoClient* conn = MongoClient::Init("127.0.0.1", 27017);
-    ASSERT_NE(nullptr, conn);
-    string gfsfilename = GetCoreFileName(newfullname);
-    MongoGridFs* gfs = new MongoGridFs(conn->GetGridFs("test", "spatial"));
-    gfs->RemoveFile(gfsfilename);
-    rs_->OutputToMongoDB(gfsfilename, gfs);
-    // Currently, the positions data of mask still has not been calculated.
-    clsRasterData<float, int>* mongors = clsRasterData<float, int>::Init(gfs, gfsfilename.c_str(), true, maskrs_, true);
-    // test mongors data
-    EXPECT_EQ(73, mongors->GetCellNumber()); // m_nCells
-    EXPECT_EQ(3, mongors->GetLayers());
-    EXPECT_EQ(64, mongors->GetValidNumber(1));
-    EXPECT_FLOAT_EQ(8.43900000f, mongors->GetAverage(3));
-    // output to asc/tif file for comparison
-    EXPECT_TRUE(mongors->OutputToFile(newfullname4mongo));
+    if (nullptr != conn) {
+        string gfsfilename = GetCoreFileName(newfullname);
+        MongoGridFs* gfs = new MongoGridFs(conn->GetGridFs("test", "spatial"));
+        gfs->RemoveFile(gfsfilename);
+        rs_->OutputToMongoDB(gfsfilename, gfs);
+        // Currently, the positions data of mask still has not been calculated.
+        clsRasterData<float, int>* mongors = clsRasterData<float, int>::Init(gfs, gfsfilename.c_str(), true, maskrs_,
+                                                                             true);
+        // test mongors data
+        EXPECT_EQ(73, mongors->GetCellNumber()); // m_nCells
+        EXPECT_EQ(3, mongors->GetLayers());
+        EXPECT_EQ(64, mongors->GetValidNumber(1));
+        EXPECT_FLOAT_EQ(8.43900000f, mongors->GetAverage(3));
+        // output to asc/tif file for comparison
+        EXPECT_TRUE(mongors->OutputToFile(newfullname4mongo));
+    }
 #endif
